[Normalizer] Combine Nearby Blocks in SeqExprs
This PR improve the normalizer, which can combine nearby binding blocks in a SeqExpr. (i.e. combine two BindingBlock or two nearby DataflowBlock)
